Rename auto-updater.js -> updater.js
To remove confusion between Electron autoUpdater and our autoUpdater
module.
